refactor(director): separar DirectorPage en vistas de mallas y avances
